admin-python/pyproject.toml
[project]
name = "admin-python"
version = "1.0.0"
description = "Admin Platform Python Backend with AI Agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "greenlet>=3.0.0",

    # Redis
    "redis[asyncio]>=5.0.0",

    # Authentication
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",

    # Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",

    # AI / LLM
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "anthropic>=0.18.0",
    "langgraph>=0.2.0",

    # MCP (Model Context Protocol)
    "mcp>=1.0.0",

    # Utilities
    "httpx>=0.26.0",
    "tenacity>=8.2.0",
    "orjson>=3.9.0",
    "gitpython>=3.1.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
